thighten compiler errors even further for playground.thibautd
unecessary cast -> error
unused import -> error
raw types -> error

git-svn-id: file:///srv/local/data/cleanRepo/trunk@22887 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/thibautd/src/main/java/playground/thibautd/initialdemandgeneration/old/activitychainsextractor/MzAdress.java
@@ -33,7 +33,7 @@ class MzAdress {
 		log.info( "number of created full adresses: "+nCreatedFullAdresses );
 		log.info( "=> proportion of adresses only consisting in ZIP Code: "+
 				(100 * ( ((double) nCreatedAdresses - (double) nCreatedFullAdresses)
-						 / (double) nCreatedAdresses ))+
+						 / nCreatedAdresses ))+
 				"%");
 	}
